requirements.txt
pandas
tantivy
agno
groq
python-dotenv
duckduckgo-search
pypdf
yfinance
uvicorn
sqlalchemy 
fastapi[standard]
